requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
Pillow>=10.2